+++ R/errorHandling.R
@@ -42,6 +42,7 @@
 #'   myFunc(x=userInput)
 #' }, silent=TRUE)
 #' stopOnError(result, "Unable to process user input")
+#'
 #' }
 
 stopOnError <- function(