+++ NEWS.md
@@ -1,5 +1,6 @@
 # MazamaCoreUtils 0.4.8
 
+* Fixed bug in `dateRange()` when creating single day date ranges.
 * Added `"ymdThms"` format to `timeStamp()`.
 * Added `"msec"` unit option to `timestamp()`.
 * Renamed default branch from 'master' to 'main'.